+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.8.1] - 2018-06-12
+### Added
+User creation logs
+
+## [2.8.0] - 2018-06-12
+### Fixed
+Social Link error message added
+Undefined index fixed
+Redirection settings added
+Short array syntax
+Added composer.json
+Redirection settings added
+Translation usage fixed
+
 ## [2.7.2] - 2018-01-18
 ### Fixed
 Block crash (class unfound) fixed